@@ -21,7 +21,7 @@ classifiers =
     Operating System :: POSIX :: Linux
     Programming Language :: Python
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.11
     Topic :: System :: Archiving
     Topic :: System :: Archiving :: Backup
     Topic :: System :: Archiving :: Compression
@@ -33,31 +33,31 @@ classifiers =
 zip_safe = False
 include_package_data = True
 packages = a3m
-python_requires = ~=3.9
+python_requires = ~=3.11
 install_requires =
     # Used by client modules
     ammcpc~=0.1
     metsrw~=0.4
-    bagit~=1.7
-    lxml~=4.7
+    bagit~=1.8
+    lxml~=4.9
     unidecode~=1.3
     pygfried~=0.6
     # Django ORM
-    Django~=3.2
+    Django~=4.2
     # Infra
-    prometheus_client~=0.12
-    requests~=2.26
+    prometheus_client~=0.17
+    requests~=2.31
     appdirs~=1.4
-    click~=8.0
-    rich~=10.16
-    tenacity~=8.0
+    click~=8.1
+    rich~=13.6
+    tenacity~=8.2
     boto3~=1.28
     # MCPServer
-    jsonschema~=4.3
-    grpcio~=1.56.2
-    grpcio-reflection~=1.56.2
-    grpcio-status~=1.56.2
-    googleapis-common-protos~=1.60.0
+    jsonschema~=4.19
+    grpcio~=1.59
+    grpcio-reflection~=1.59
+    grpcio-status~=1.59
+    googleapis-common-protos~=1.61
 
 
 [options.extras_require]
@@ -67,7 +67,6 @@ dev =
     pytest-django
     pytest-env
     pytest-mock
-    vcrpy
     coverage
     pip-tools
     grpcio-tools
