@@ -1,4 +1,4 @@
-# Generated by Django 3.2.20 on 2023-10-20 12:51
+# Generated by Django 4.2.6 on 2023-10-23 20:02
 import uuid
 
 import django.db.models.deletion
@@ -64,6 +64,152 @@ class Migration(migrations.Migration):
             ],
             options={
                 "db_table": "Agents",
+            },
+        ),
+        migrations.CreateModel(
+            name="Derivation",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(
+                        db_column="pk",
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+            ],
+            options={
+                "db_table": "Derivations",
+            },
+        ),
+        migrations.CreateModel(
+            name="Directory",
+            fields=[
+                (
+                    "uuid",
+                    models.CharField(
+                        db_column="directoryUUID",
+                        max_length=36,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                (
+                    "originallocation",
+                    a3m.main.models.BlobTextField(db_column="originalLocation"),
+                ),
+                (
+                    "currentlocation",
+                    a3m.main.models.BlobTextField(
+                        db_column="currentLocation", null=True
+                    ),
+                ),
+                (
+                    "enteredsystem",
+                    models.DateTimeField(auto_now_add=True, db_column="enteredSystem"),
+                ),
+            ],
+            options={
+                "db_table": "Directories",
+            },
+        ),
+        migrations.CreateModel(
+            name="DublinCore",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(db_column="pk", primary_key=True, serialize=False),
+                ),
+                (
+                    "metadataappliestoidentifier",
+                    models.CharField(
+                        blank=True,
+                        db_column="metadataAppliesToidentifier",
+                        default=None,
+                        max_length=36,
+                        null=True,
+                    ),
+                ),
+                ("title", models.TextField(blank=True, db_column="title")),
+                ("creator", models.TextField(blank=True, db_column="creator")),
+                ("subject", models.TextField(blank=True, db_column="subject")),
+                ("description", models.TextField(blank=True, db_column="description")),
+                ("publisher", models.TextField(blank=True, db_column="publisher")),
+                ("contributor", models.TextField(blank=True, db_column="contributor")),
+                (
+                    "date",
+                    models.TextField(
+                        blank=True,
+                        db_column="date",
+                        help_text="Use ISO 8601 (YYYY-MM-DD or YYYY-MM-DD/YYYY-MM-DD)",
+                    ),
+                ),
+                ("type", models.TextField(blank=True, db_column="type")),
+                ("format", models.TextField(blank=True, db_column="format")),
+                ("identifier", models.TextField(blank=True, db_column="identifier")),
+                ("source", models.TextField(blank=True, db_column="source")),
+                ("relation", models.TextField(blank=True, db_column="relation")),
+                (
+                    "language",
+                    models.TextField(
+                        blank=True, db_column="language", help_text="Use ISO 639"
+                    ),
+                ),
+                ("coverage", models.TextField(blank=True, db_column="coverage")),
+                ("rights", models.TextField(blank=True, db_column="rights")),
+                (
+                    "status",
+                    models.CharField(
+                        choices=[("ORIGINAL", "original"), ("UPDATED", "updated")],
+                        db_column="status",
+                        default="ORIGINAL",
+                        max_length=8,
+                    ),
+                ),
+            ],
+            options={
+                "db_table": "Dublincore",
+            },
+        ),
+        migrations.CreateModel(
+            name="Event",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(
+                        db_column="pk",
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                (
+                    "event_id",
+                    models.UUIDField(
+                        db_column="eventIdentifierUUID",
+                        default=uuid.uuid4,
+                        null=True,
+                        unique=True,
+                    ),
+                ),
+                ("event_type", models.TextField(blank=True, db_column="eventType")),
+                (
+                    "event_datetime",
+                    models.DateTimeField(auto_now=True, db_column="eventDateTime"),
+                ),
+                ("event_detail", models.TextField(blank=True, db_column="eventDetail")),
+                (
+                    "event_outcome",
+                    models.TextField(blank=True, db_column="eventOutcome"),
+                ),
+                (
+                    "event_outcome_detail",
+                    models.TextField(blank=True, db_column="eventOutcomeDetailNote"),
+                ),
+            ],
+            options={
+                "db_table": "Events",
             },
         ),
         migrations.CreateModel(
@@ -135,6 +281,65 @@ class Migration(migrations.Migration):
             options={
                 "db_table": "Files",
             },
+        ),
+        migrations.CreateModel(
+            name="FileFormatVersion",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(
+                        db_column="pk",
+                        editable=False,
+                        primary_key=True,
+                        serialize=False,
+                    ),
+                ),
+                ("format_version_id", models.UUIDField(editable=False)),
+            ],
+            options={
+                "db_table": "FilesIdentifiedIDs",
+            },
+        ),
+        migrations.CreateModel(
+            name="FileID",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(db_column="pk", primary_key=True, serialize=False),
+                ),
+                ("format_name", models.TextField(blank=True, db_column="formatName")),
+                (
+                    "format_version",
+                    models.TextField(blank=True, db_column="formatVersion"),
+                ),
+                (
+                    "format_registry_name",
+                    models.TextField(blank=True, db_column="formatRegistryName"),
+                ),
+                (
+                    "format_registry_key",
+                    models.TextField(blank=True, db_column="formatRegistryKey"),
+                ),
+            ],
+            options={
+                "db_table": "FilesIDs",
+            },
+        ),
+        migrations.CreateModel(
+            name="FPCommandOutput",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(
+                        auto_created=True,
+                        primary_key=True,
+                        serialize=False,
+                        verbose_name="ID",
+                    ),
+                ),
+                ("content", models.TextField(null=True)),
+                ("rule_id", models.UUIDField(editable=False)),
+            ],
         ),
         migrations.CreateModel(
             name="Identifier",
@@ -229,17 +434,6 @@ class Migration(migrations.Migration):
             ],
             options={
                 "db_table": "Jobs",
-                "index_together": {
-                    ("jobtype", "currentstep"),
-                    (
-                        "sipuuid",
-                        "currentstep",
-                        "microservicegroup",
-                        "microservicechainlink",
-                    ),
-                    ("sipuuid", "createdtime", "createdtimedec"),
-                    ("sipuuid", "jobtype", "createdtime", "createdtimedec"),
-                },
             },
         ),
         migrations.CreateModel(
@@ -954,7 +1148,7 @@ class Migration(migrations.Migration):
                     models.TextField(blank=True, db_column="aipFilename", null=True),
                 ),
                 ("diruuids", models.BooleanField(db_column="dirUUIDs", default=False)),
-                ("identifiers", models.ManyToManyField(to="main.Identifier")),
+                ("identifiers", models.ManyToManyField(to="main.identifier")),
             ],
             options={
                 "db_table": "SIPs",
@@ -1367,95 +1561,71 @@ class Migration(migrations.Migration):
                 "db_table": "RightsStatementCopyrightDocumentationIdentifier",
             },
         ),
-        migrations.CreateModel(
-            name="FPCommandOutput",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(
-                        auto_created=True,
-                        primary_key=True,
-                        serialize=False,
-                        verbose_name="ID",
-                    ),
-                ),
-                ("content", models.TextField(null=True)),
-                ("rule_id", models.UUIDField(editable=False)),
-                (
-                    "file",
-                    models.ForeignKey(
-                        db_column="fileUUID",
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.file",
-                    ),
-                ),
-            ],
+        migrations.AddIndex(
+            model_name="job",
+            index=models.Index(
+                fields=["sipuuid", "createdtime", "createdtimedec"],
+                name="Jobs_SIPUUID_6cf6c1_idx",
+            ),
         ),
-        migrations.CreateModel(
-            name="FileID",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(db_column="pk", primary_key=True, serialize=False),
-                ),
-                ("format_name", models.TextField(blank=True, db_column="formatName")),
-                (
-                    "format_version",
-                    models.TextField(blank=True, db_column="formatVersion"),
-                ),
-                (
-                    "format_registry_name",
-                    models.TextField(blank=True, db_column="formatRegistryName"),
-                ),
-                (
-                    "format_registry_key",
-                    models.TextField(blank=True, db_column="formatRegistryKey"),
-                ),
-                (
-                    "file",
-                    models.ForeignKey(
-                        blank=True,
-                        db_column="fileUUID",
-                        null=True,
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.file",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "FilesIDs",
-            },
+        migrations.AddIndex(
+            model_name="job",
+            index=models.Index(
+                fields=["sipuuid", "jobtype", "createdtime", "createdtimedec"],
+                name="Jobs_SIPUUID_cf4b11_idx",
+            ),
         ),
-        migrations.CreateModel(
-            name="FileFormatVersion",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(
-                        db_column="pk",
-                        editable=False,
-                        primary_key=True,
-                        serialize=False,
-                    ),
-                ),
-                ("format_version_id", models.UUIDField(editable=False)),
-                (
-                    "file_uuid",
-                    models.ForeignKey(
-                        db_column="fileUUID",
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.file",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "FilesIdentifiedIDs",
-            },
+        migrations.AddIndex(
+            model_name="job",
+            index=models.Index(
+                fields=[
+                    "sipuuid",
+                    "currentstep",
+                    "microservicegroup",
+                    "microservicechainlink",
+                ],
+                name="Jobs_SIPUUID_658a37_idx",
+            ),
+        ),
+        migrations.AddIndex(
+            model_name="job",
+            index=models.Index(
+                fields=["jobtype", "currentstep"], name="Jobs_jobType_4a3346_idx"
+            ),
+        ),
+        migrations.AddField(
+            model_name="fpcommandoutput",
+            name="file",
+            field=models.ForeignKey(
+                db_column="fileUUID",
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.file",
+            ),
+        ),
+        migrations.AddField(
+            model_name="fileid",
+            name="file",
+            field=models.ForeignKey(
+                blank=True,
+                db_column="fileUUID",
+                null=True,
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.file",
+            ),
+        ),
+        migrations.AddField(
+            model_name="fileformatversion",
+            name="file_uuid",
+            field=models.ForeignKey(
+                db_column="fileUUID",
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.file",
+            ),
         ),
         migrations.AddField(
             model_name="file",
             name="identifiers",
-            field=models.ManyToManyField(to="main.Identifier"),
+            field=models.ManyToManyField(to="main.identifier"),
         ),
         migrations.AddField(
             model_name="file",
@@ -1479,223 +1649,94 @@ class Migration(migrations.Migration):
                 to="main.transfer",
             ),
         ),
-        migrations.CreateModel(
-            name="Event",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(
-                        db_column="pk",
-                        editable=False,
-                        primary_key=True,
-                        serialize=False,
-                    ),
-                ),
-                (
-                    "event_id",
-                    models.UUIDField(
-                        db_column="eventIdentifierUUID",
-                        default=uuid.uuid4,
-                        null=True,
-                        unique=True,
-                    ),
-                ),
-                ("event_type", models.TextField(blank=True, db_column="eventType")),
-                (
-                    "event_datetime",
-                    models.DateTimeField(auto_now=True, db_column="eventDateTime"),
-                ),
-                ("event_detail", models.TextField(blank=True, db_column="eventDetail")),
-                (
-                    "event_outcome",
-                    models.TextField(blank=True, db_column="eventOutcome"),
-                ),
-                (
-                    "event_outcome_detail",
-                    models.TextField(blank=True, db_column="eventOutcomeDetailNote"),
-                ),
-                ("agents", models.ManyToManyField(to="main.Agent")),
-                (
-                    "file_uuid",
-                    models.ForeignKey(
-                        blank=True,
-                        db_column="fileUUID",
-                        null=True,
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.file",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "Events",
-            },
+        migrations.AddField(
+            model_name="event",
+            name="agents",
+            field=models.ManyToManyField(to="main.agent"),
         ),
-        migrations.CreateModel(
-            name="DublinCore",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(db_column="pk", primary_key=True, serialize=False),
-                ),
-                (
-                    "metadataappliestoidentifier",
-                    models.CharField(
-                        blank=True,
-                        db_column="metadataAppliesToidentifier",
-                        default=None,
-                        max_length=36,
-                        null=True,
-                    ),
-                ),
-                ("title", models.TextField(blank=True, db_column="title")),
-                ("creator", models.TextField(blank=True, db_column="creator")),
-                ("subject", models.TextField(blank=True, db_column="subject")),
-                ("description", models.TextField(blank=True, db_column="description")),
-                ("publisher", models.TextField(blank=True, db_column="publisher")),
-                ("contributor", models.TextField(blank=True, db_column="contributor")),
-                (
-                    "date",
-                    models.TextField(
-                        blank=True,
-                        db_column="date",
-                        help_text="Use ISO 8601 (YYYY-MM-DD or YYYY-MM-DD/YYYY-MM-DD)",
-                    ),
-                ),
-                ("type", models.TextField(blank=True, db_column="type")),
-                ("format", models.TextField(blank=True, db_column="format")),
-                ("identifier", models.TextField(blank=True, db_column="identifier")),
-                ("source", models.TextField(blank=True, db_column="source")),
-                ("relation", models.TextField(blank=True, db_column="relation")),
-                (
-                    "language",
-                    models.TextField(
-                        blank=True, db_column="language", help_text="Use ISO 639"
-                    ),
-                ),
-                ("coverage", models.TextField(blank=True, db_column="coverage")),
-                ("rights", models.TextField(blank=True, db_column="rights")),
-                (
-                    "status",
-                    models.CharField(
-                        choices=[("ORIGINAL", "original"), ("UPDATED", "updated")],
-                        db_column="status",
-                        default="ORIGINAL",
-                        max_length=8,
-                    ),
-                ),
-                (
-                    "metadataappliestotype",
-                    models.ForeignKey(
-                        db_column="metadataAppliesToType",
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.metadataappliestotype",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "Dublincore",
-            },
+        migrations.AddField(
+            model_name="event",
+            name="file_uuid",
+            field=models.ForeignKey(
+                blank=True,
+                db_column="fileUUID",
+                null=True,
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.file",
+            ),
         ),
-        migrations.CreateModel(
-            name="Directory",
-            fields=[
-                (
-                    "uuid",
-                    models.CharField(
-                        db_column="directoryUUID",
-                        max_length=36,
-                        primary_key=True,
-                        serialize=False,
-                    ),
-                ),
-                (
-                    "originallocation",
-                    a3m.main.models.BlobTextField(db_column="originalLocation"),
-                ),
-                (
-                    "currentlocation",
-                    a3m.main.models.BlobTextField(
-                        db_column="currentLocation", null=True
-                    ),
-                ),
-                (
-                    "enteredsystem",
-                    models.DateTimeField(auto_now_add=True, db_column="enteredSystem"),
-                ),
-                ("identifiers", models.ManyToManyField(to="main.Identifier")),
-                (
-                    "sip",
-                    models.ForeignKey(
-                        blank=True,
-                        db_column="sipUUID",
-                        null=True,
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.sip",
-                    ),
-                ),
-                (
-                    "transfer",
-                    models.ForeignKey(
-                        blank=True,
-                        db_column="transferUUID",
-                        null=True,
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.transfer",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "Directories",
-            },
+        migrations.AddField(
+            model_name="dublincore",
+            name="metadataappliestotype",
+            field=models.ForeignKey(
+                db_column="metadataAppliesToType",
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.metadataappliestotype",
+            ),
         ),
-        migrations.CreateModel(
-            name="Derivation",
-            fields=[
-                (
-                    "id",
-                    models.AutoField(
-                        db_column="pk",
-                        editable=False,
-                        primary_key=True,
-                        serialize=False,
-                    ),
-                ),
-                (
-                    "derived_file",
-                    models.ForeignKey(
-                        db_column="derivedFileUUID",
-                        on_delete=django.db.models.deletion.CASCADE,
-                        related_name="original_file_set",
-                        to="main.file",
-                    ),
-                ),
-                (
-                    "event",
-                    models.ForeignKey(
-                        blank=True,
-                        db_column="relatedEventUUID",
-                        null=True,
-                        on_delete=django.db.models.deletion.CASCADE,
-                        to="main.event",
-                        to_field="event_id",
-                    ),
-                ),
-                (
-                    "source_file",
-                    models.ForeignKey(
-                        db_column="sourceFileUUID",
-                        on_delete=django.db.models.deletion.CASCADE,
-                        related_name="derived_file_set",
-                        to="main.file",
-                    ),
-                ),
-            ],
-            options={
-                "db_table": "Derivations",
-            },
+        migrations.AddField(
+            model_name="directory",
+            name="identifiers",
+            field=models.ManyToManyField(to="main.identifier"),
         ),
-        migrations.AlterIndexTogether(
-            name="file",
-            index_together={("sip", "filegrpuse")},
+        migrations.AddField(
+            model_name="directory",
+            name="sip",
+            field=models.ForeignKey(
+                blank=True,
+                db_column="sipUUID",
+                null=True,
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.sip",
+            ),
+        ),
+        migrations.AddField(
+            model_name="directory",
+            name="transfer",
+            field=models.ForeignKey(
+                blank=True,
+                db_column="transferUUID",
+                null=True,
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.transfer",
+            ),
+        ),
+        migrations.AddField(
+            model_name="derivation",
+            name="derived_file",
+            field=models.ForeignKey(
+                db_column="derivedFileUUID",
+                on_delete=django.db.models.deletion.CASCADE,
+                related_name="original_file_set",
+                to="main.file",
+            ),
+        ),
+        migrations.AddField(
+            model_name="derivation",
+            name="event",
+            field=models.ForeignKey(
+                blank=True,
+                db_column="relatedEventUUID",
+                null=True,
+                on_delete=django.db.models.deletion.CASCADE,
+                to="main.event",
+                to_field="event_id",
+            ),
+        ),
+        migrations.AddField(
+            model_name="derivation",
+            name="source_file",
+            field=models.ForeignKey(
+                db_column="sourceFileUUID",
+                on_delete=django.db.models.deletion.CASCADE,
+                related_name="derived_file_set",
+                to="main.file",
+            ),
+        ),
+        migrations.AddIndex(
+            model_name="file",
+            index=models.Index(
+                fields=["sip", "filegrpuse"], name="Files_sipUUID_a0d05d_idx"
+            ),
         ),
     ]
